+++ packages/react-icons-atomic-webpack-loader/src/loader-context.ts
@@ -13,6 +13,7 @@ import type { FluentIconsAtomicImportLoaderOptions } from './index';
 export interface AtomicLoaderContext {
   readonly resourcePath: string;
   readonly resourceQuery: string;
+  readonly sourceMap?: boolean;
   readonly _compilation?: unknown;
   getOptions(): FluentIconsAtomicImportLoaderOptions;
   callback(err: Error | null | undefined, content?: string | Buffer, sourceMap?: any, additionalData?: any): void;