+++ packages/react-icons-svg-sprite-subsetting-webpack-plugin/test/src/multi-entry-backpack-lazy.js
@@ -1,0 +1,3 @@
+import { BackpackFilled } from '@fluentui/react-icons/svg-sprite/backpack';
+
+export { BackpackFilled };