+++ packages/react-icons-svg-sprite-subsetting-webpack-plugin/test/src/multi-entry-backpack.js
@@ -1,0 +1,1 @@
+import('./multi-entry-backpack-lazy.js').then(({ BackpackFilled }) => console.dir({ BackpackFilled }));